Rewind back to the time when COVID-19 had just struck.. The next challenge was to bring these vaccines to the people and vaccinate all of citizens of the country starting with the most vulnerable.
Introduction
Rewind back to the time when COVID-19 had just struck. The government had to procure
vaccines from different countries and labs. The next challenge was to bring these vaccines to
the people and vaccinate all of citizens of the country starting with the most vulnerable.
This setup was massive and the government had to act swiftly to ensure the safety of the
citizens.
You are required to create a system that would handle and help the government in this process.
Modules
If we break down the whole process in chunks it would come down to the following modules:
Super Admin
The super admin can perform CRUD (Create, Retrieve, Update and Delete) operation on all user
accounts.
Admin
Procuring vaccines from companies would be the first step. These can be both local and
international purchases.
We’d then have to store these vaccines in government owned warehouse facilities spread
across the country.
The next task would be to establish vaccination centers spread across the country. The
government would have to ensure that each city gets at least one vaccination center. The
population to vaccination center ratio must not be any more than 1:20,000 i.e. for every 20,000
citizens there should be at least one vaccination center. This number can be kept dynamic.
Now the vaccinations have to be sent to the vaccination centers daily depending on the
consumption of that vaccination center.
Doctor
Doctors would be assigned to each vaccination center along with their duty days. Doctors are
required to assess measure the citizen’s blood pressure, oxygen level, and glucose level. The

National University of Computer & Emerging Sciences – FAST (CFD)
Department of Computer Science: Object Oriented Programming - Lab

P a g e 4 | 7
doctors would just retrieve citizen information from the file and log in their vital stats along
with the date.
Front Desk Officer (FDO)
The FDO would be the first person the citizen would interact with. There are multiple counters
in the vaccination center. On each counter an MD Doctor would checkup the patient, and then
administer a vaccine from the ones available in stock at the vaccination center.
FDO would check citizen eligibility, if eligible the system would assign a counter to the citizen
that is the least busy.
Government Official
Government officials would be able to see reports of the vaccination drive in the country and it
would include various statistics. Including but not limited to:
 Percentage of the vaccinated population
 City wise vaccine data
 Percentage of district vaccinated population
 Amount spent on vaccines so far
 Filter only those patients who took a specific type of vaccine.
 Filter patients who are partially vaccinated.
 Filter patients who are fully vaccinated.
 Filter patients who took the booster shot.
 View a statistical tabular graph showing the percentages of how many people are
partially vaccinated, how many are fully vaccinated, and how many took booster.

National University of Computer & Emerging Sciences – FAST (CFD)
Department of Computer Science: Object Oriented Programming - Lab

P a g e 5 | 7

Citizen
The individuals will go through the following process.
Step 1: Sign up/Log in:
All the citizens have to do is sign up using a unique email and password that has not yet been
taken by another citizen also registered on the portal. Make a proper sign up and log in page for
any individual who wishes to register him or herself in order to get the vaccine.
On the sign up form:
Ask for:
 First Name
 Last Name
 Phone Number
 Email
 CNIC
 Blood Type
 City
 State
 Address
 DOB
(this information will also go in the attributes of RegisteredCitizen class. Rest will be filled by
administration)
Step 2: Check Eligibility Criteria
While filling out the form, the user must be asked if he has any history of severe allergic
reactions to any of the components in the vaccine. If yes, he cannot register and the
eligibilityCheck attribute will be set to False.
Secondly, user’s age must be 5 or above.
Step 3: Receive Confirmation
Self-explanatory. Display a nice format for confirmation notification.
Lastly...
The patient should be able to view his/her history.

National University of Computer & Emerging Sciences – FAST (CFD)
Department of Computer Science: Object Oriented Programming - Lab

P a g e 6 | 7

 Review a record of which vaccines they got, on which dates.
 Review the date for which their next vaccine is scheduled.
Requirements
All users would be required to sign in using a username and password. These login credentials
would be stored on .txt files on the system and the passwords would be encrypted.
Show sign in page to each type of user and ask for username and password. The username and
password should be unique for each type of member.
Information required for vaccines:
 vacType (WHOLE VIRUS, PROTEIN SUBUNIT, VIRAL VECTOR and NUCLEIC ACID
 vacName (sinopharm, pfizer)
 vacID
 Number of doses
 Price
 Batch ID
Vaccines would be stored in warehouses, batches would be purchases and stored. Expired
vaccines must be given higher priority when dispatching to vaccine centers.
Information required for vaccines:
 Warehouse name
 Warehouse location (latitude, longitude)
 Warehouse capacity
Information required per batch
 Vaccine ID
 suppliedById (got from which supplier? His ID) [must be valid]
 Expiry date

All items would be given a system generated ID, for instance when adding a new vaccine a new
ID would be generated and assigned to the vaccine. You may use .csv files or .txt to store data.
NOTE: Every class involving a person as an entity such as manufacturer, vendor, supplier, FDO
or civilian should be inherited from a parent class “PERSON” which has the attributes that are
COMMON among all the child classes such as firstName, lastName, email, contactNumber and
ID.